Make config's fields privately and fix shutdown problem in DB and add StatusService in KafkaAdminController
각 브랜치를 merge하는 과정 중 누락된 KafkaAdminController 에서의 StatusService 메소드들을 다시 추가하였다.
그리고 컨트롤러에서의 configs shutdown 과정중 결과가 DB에 제대로 반영되지않는 문제를 해결하였다.
잘못된 JSON 데이터를 받았을 때 그것을 정상적으로 처리해주는 method 를 ConsumerBufferConfigs에 구현해야한다.

추가) Change configs 시에 argument에 wrong JSON 이 들어온지를 검사해주는 메소드를 구현함.

+++ Apiserver/src/main/java/kr/ac/hongik/apl/broker/apiserver/Controller/KafkaAdminController.java
@@ -83,6 +83,7 @@ public class KafkaAdminController {
             if (!consumerDataService.checkTopic(topicName)) {
                 BufferedConsumingPbftClient bufferedConsumingPbftClient =
                         consumerFactoryService.MakeBufferedConsumer(configs.getCommonConfigs(), configs.getBuffConfigs());
+                statusService.addBufferStatus(configs);
                 consumerDataService.setConsumer(topicName,bufferedConsumingPbftClient);
                 asyncExecutionService.runAsConsumerExecutor(bufferedConsumingPbftClient::startConsumer);
                 // TODO : client 객체 db에 삽입.
@@ -104,6 +105,7 @@ public class KafkaAdminController {
             if (!consumerDataService.checkTopic(topicName)) {
                 ImmediateConsumingPbftClient immediateConsumingPbftClient =
                         consumerFactoryService.MakeImmediateConsumer(configs.getCommonConfigs(), configs.getImmeConfigs());
+                statusService.addImmediateStatus(configs);
                 consumerDataService.setConsumer(topicName,immediateConsumingPbftClient);
                 asyncExecutionService.runAsConsumerExecutor(immediateConsumingPbftClient::startConsumer);
                 return String.format("New Immediate Consumer added!");
@@ -136,14 +138,23 @@ public class KafkaAdminController {
     @PostMapping("/consumer/change/buf")
     @ResponseBody
     public String changeBuffConsumer(@RequestBody ConsumerBufferConfigs configs) throws Exception {
+
+        log.info(configs.toString());
+        if(configs.checkFieldsNull())
+        {
+            return String.format("you've missed some of fields. try again");
+        }
+
         int topicListSize = configs.getBuffTopicName().size();
         if (topicListSize == 1) {
             String topicName = configs.getBuffTopicName().get(0);
             if (consumerDataService.checkTopic(topicName)) {
+                statusService.deleteBufferStatus(configs);
                 consumerDataService.consumerShutdown(topicName);
                 consumerDataService.deleteData(topicName);
                 BufferedConsumingPbftClient bufferedConsumingPbftClient =
                         consumerFactoryService.MakeBufferedConsumer(configs.getCommonConfigs(), configs.getBuffConfigs());
+                statusService.addBufferStatus(configs);
                 consumerDataService.setConsumer(topicName,bufferedConsumingPbftClient);
                 asyncExecutionService.runAsConsumerExecutor(bufferedConsumingPbftClient::startConsumer);
                 log.info("Buffer Consumer has been changed");
@@ -216,6 +227,7 @@ public class KafkaAdminController {
     @ResponseBody
     public String shutdownBuffer(@RequestParam(value = "topicName", required = true, defaultValue = "") String topicName) throws Exception {
         if (consumerDataService.checkTopic(topicName)) {
+            statusService.deleteBufferStatus(topicName);
             consumerDataService.consumerShutdown(topicName);
             consumerDataService.deleteData(topicName);
             return "Request - shutdown consumer : " + topicName;
